Fix issue when there are lot of tasks too be dequeued
If there are more that 60k tasks that needs to be dequeued, queue.pull
raise pg exception of out_of_memeory, reason is it tries to acquire
advisory_xact_lock lock for each row.

But with max_locks_per_transaction set to 100 it fails which is already
a big number, based on number of transactions we have.

Only take lock on 1st thousand instead of all records

[Z-0000]